requirements.txt
Pillow==9.0.1
requests==2.27.1